# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  controller.CreatePostRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
...
  controller.SwaggerCategoryResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: 技术相关文章
        type: string
      id:
        example: 1
        type: integer
      name:
        example: 技术
        type: string
      post_count:
        example: 10
        type: integer
      slug:
        example: ji-shu
        type: string
      sort_order:
        example: 1
        type: integer
    type: object
...
  controller.SwaggerPostListData:
    properties:
      list:
        items:
          $ref: '#/definitions/controller.SwaggerPostResponse'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  controller.SwaggerPostResponse:
    properties:
      category:
        $ref: '#/definitions/controller.SwaggerCategoryResponse'
      category_id:
        example: 1
        type: integer
      comment_count:
        example: 5
        type: integer
      content:
        example: |-
          # 标题

          内容...
        type: string
      cover_image:
        example: https://example.com/cover.jpg
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      is_top:
        example: false
        type: boolean
      like_count:
        example: 10
        type: integer
      slug:
        example: go-language-intro
        type: string
      status:
        example: 1
        type: integer
      summary:
        example: 文章摘要
        type: string
      tags:
        items:
          $ref: '#/definitions/controller.SwaggerTagResponse'
        type: array
      title:
        example: Go语言入门
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/controller.SwaggerUserResponse'
      user_id:
        example: 1
        type: integer
      view_count:
        example: 100
        type: integer
    type: object
...
  controller.SwaggerTagResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Go
        type: string
      post_count:
        example: 5
        type: integer
      slug:
        example: go
        type: string
    type: object
  controller.SwaggerUserResponse:
    properties:
      avatar:
        example: https://example.com/avatar.jpg
        type: string
      bio:
        example: 个人简介
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: test@example.com
        type: string
      id:
        example: 1
        type: integer
      role:
        example: user
        type: string
      username:
        example: testuser
        type: string
    type: object
...
  controller.UpdatePostRequest:
    properties:
      category_id:
        type: integer
      content:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        maxLength: 200
        minLength: 1
        type: string
    required:
    - content
    - title
    type: object
...
  utils.Response:
    properties:
      code:
        description: 状态码，0表示成功
        example: 0
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        example: success
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: your-email@example.com
    name: API Support
  description: 个人博客系统 API 文档
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Blog API
  version: "1.0"
paths:
...
  /posts:
    get:
      consumes:
      - application/json
      description: 分页获取文章列表
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - default: 1
        description: 状态 1:已发布 0:草稿 2:私密
        in: query
        name: status
        type: integer
      - description: 分类ID
        in: query
        name: category_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerPostListData'
              type: object
      summary: 获取文章列表
      tags:
      - 文章管理
    post:
      consumes:
      - application/json
      description: 创建新文章（需要认证）
      parameters:
      - description: 文章信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.CreatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerPostResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 创建文章
      tags:
      - 文章管理
  /posts/{id}:
    delete:
      consumes:
      - application/json
      description: 删除文章（需要认证，仅作者或管理员）
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 删除文章
      tags:
      - 文章管理
    get:
      consumes:
      - application/json
      description: 根据ID获取文章详情
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerPostResponse'
              type: object
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 获取文章详情
      tags:
      - 文章管理
    put:
      consumes:
      - application/json
      description: 更新文章内容（需要认证，仅作者或管理员）
      parameters:
      - description: 文章ID
        in: path
        name: id
        required: true
        type: integer
      - description: 文章信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.UpdatePostRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerPostResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/utils.Response'
        "401":
          description: 未授权
          schema:
            $ref: '#/definitions/utils.Response'
        "403":
          description: 权限不足
          schema:
            $ref: '#/definitions/utils.Response'
        "404":
          description: 文章不存在
          schema:
            $ref: '#/definitions/utils.Response'
      security:
      - BearerAuth: []
      summary: 更新文章
      tags:
      - 文章管理
...
  /posts/tag/{slug}:
    get:
      consumes:
      - application/json
      description: 通过标签slug获取该标签下的所有文章
      parameters:
      - description: 标签slug
        in: path
        name: slug
        required: true
        type: string
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/utils.Response'
            - properties:
                data:
                  $ref: '#/definitions/controller.SwaggerPostListData'
              type: object
        "404":
          description: 标签不存在
          schema:
            $ref: '#/definitions/utils.Response'
      summary: 根据标签获取文章
      tags:
      - 文章管理
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"